feat: 기업 상세 조회 추가

See merge request s12-final/S12P31B105!113

+++ backend/src/main/java/com/ssafy/hellojob/domain/company/entity/Company.java
@@ -15,7 +15,7 @@ public class Company {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "company_id", nullable = false)
-    private Long companyId;
+    private Integer companyId;
 
     @Column(name = "company_name", nullable = false)
     private String companyName;
