+++ backend/src/main/java/com/ssafy/hellojob/domain/schedule/controller/ScheduleController.java
@@ -3,7 +3,6 @@ package com.ssafy.hellojob.domain.schedule.controller;
 import com.ssafy.hellojob.domain.schedule.dto.request.ScheduleAddRequestDto;
 import com.ssafy.hellojob.domain.schedule.dto.request.ScheduleUpdateScheduleCoverLetterRequestDto;
 import com.ssafy.hellojob.domain.schedule.dto.request.ScheduleUpdateScheduleStatusRequestDto;
-import com.ssafy.hellojob.domain.schedule.dto.response.ScheduleDetailCompanyAnalysis;
 import com.ssafy.hellojob.domain.schedule.dto.response.ScheduleDetailResponseDto;
 import com.ssafy.hellojob.domain.schedule.dto.response.ScheduleIdResponseDto;
 import com.ssafy.hellojob.domain.schedule.dto.response.ScheduleListResponseDto;
@@ -11,7 +10,6 @@ import com.ssafy.hellojob.domain.schedule.service.ScheduleService;
 import com.ssafy.hellojob.global.auth.token.UserPrincipal;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
@@ -39,7 +37,7 @@ public class ScheduleController {
 
     // 일정 삭제
     @DeleteMapping("{scheduleId}")
-    public void ScheduleDelete(@PathVariable("scheduleId") Long scheduleId,
+    public void ScheduleDelete(@PathVariable("scheduleId") Integer scheduleId,
                                @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
 
@@ -50,7 +48,7 @@ public class ScheduleController {
     // 일정 상태 수정
     @PatchMapping("/{scheduleId}/status")
     public ScheduleIdResponseDto ScheduleUpdateStatus(@Valid @RequestBody ScheduleUpdateScheduleStatusRequestDto requestDto,
-                                                      @PathVariable("scheduleId") Long scheduleId,
+                                                      @PathVariable("scheduleId") Integer scheduleId,
                                                       @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
 
@@ -61,7 +59,7 @@ public class ScheduleController {
     // 일정 자기소개서 수정
     @PatchMapping("/{scheduleId}/cover-letter")
     public ScheduleIdResponseDto ScheduleUpdateCoverLetter(@Valid @RequestBody ScheduleUpdateScheduleCoverLetterRequestDto requestDto,
-                                                            @PathVariable("scheduleId") Long scheduleId,
+                                                            @PathVariable("scheduleId") Integer scheduleId,
                                                             @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
 
@@ -72,7 +70,7 @@ public class ScheduleController {
     // 일정 전체 수정
     @PutMapping("/{scheduleId}")
     public ScheduleIdResponseDto ScheduleUpdate(@Valid @RequestBody ScheduleAddRequestDto requestDto,
-                                                @PathVariable("scheduleId") Long scheduleId,
+                                                @PathVariable("scheduleId") Integer scheduleId,
                                                 @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
         ScheduleIdResponseDto responseDto = scheduleService.updateSchedule(requestDto, scheduleId, userId);
@@ -90,7 +88,7 @@ public class ScheduleController {
 
     // 일정 상세 조회
     @GetMapping("/{scheduleId}")
-    public ScheduleDetailResponseDto ScheduleDetail(@PathVariable("scheduleId") Long scheduleId,
+    public ScheduleDetailResponseDto ScheduleDetail(@PathVariable("scheduleId") Integer scheduleId,
                                                     @AuthenticationPrincipal UserPrincipal userPrincipal){
         Integer userId = userPrincipal.getUserId();
         ScheduleDetailResponseDto responseDto = scheduleService.detailSchedule(scheduleId, userId);
